+++ modules/base/include/internal/static.h
@@ -11,10 +11,11 @@
 #include <IMP/base/base_config.h>
 #include <IMP/base/Vector.h>
 #include <IMP/base/map.h>
-#include <IMP/base/set.h>
+#include <IMP/base/Flag.h>
 #include <IMP/base/tuple_macros.h>
 #include <boost/cstdint.hpp>
 #include <boost/program_options.hpp>
+#include <boost/unordered_set.hpp>
 #include <boost/progress.hpp>
 #include <boost/scoped_ptr.hpp>
 
@@ -26,10 +27,12 @@ extern IMPBASEEXPORT base::map<std::string, unsigned int> object_type_counts;
 IMPBASEEXPORT void check_live_objects();
 
 extern IMPBASEEXPORT boost::program_options::options_description flags;
+extern IMPBASEEXPORT boost::program_options::options_description advanced_flags;
 extern IMPBASEEXPORT boost::program_options::variables_map variables_map;
 
-extern IMPBASEEXPORT boost::int64_t check_level;
-extern IMPBASEEXPORT boost::int64_t stats_level;
+extern IMPBASEEXPORT AdvancedFlag<CheckLevel, IMP_HAS_CHECKS != IMP_NONE>
+    check_level;
+extern IMPBASEEXPORT AdvancedFlag<StatisticsLevel> stats_level;
 #if !IMP_BASE_HAS_LOG4CXX
 extern IMPBASEEXPORT base::Vector<std::pair<const char *, const void *> >
     log_contexts;
@@ -39,32 +42,44 @@ class LogStream;
 extern IMPBASEEXPORT LogStream stream;
 #endif
 // needed for flags
-extern IMPBASEEXPORT boost::int64_t log_level;
+extern IMPBASEEXPORT Flag<LogLevel, IMP_HAS_LOG != IMP_SILENT> log_level;
 
 // Should be uint, but no flags support for that.
-extern IMPBASEEXPORT boost::int64_t random_seed;
+extern IMPBASEEXPORT AdvancedFlag<boost::int64_t> random_seed;
 
-extern IMPBASEEXPORT bool cpu_profile;
-extern IMPBASEEXPORT bool heap_profile;
+extern IMPBASEEXPORT AdvancedFlag<bool, IMP_BASE_HAS_GPERFTOOLS> cpu_profile;
+extern IMPBASEEXPORT AdvancedFlag<bool, IMP_BASE_HAS_GPERFTOOLS> heap_profile;
+extern IMPBASEEXPORT Flag<bool> help;
+extern IMPBASEEXPORT Flag<bool> help_advanced;
+extern IMPBASEEXPORT Flag<bool> version;
+extern IMPBASEEXPORT AdvancedFlag<bool> show_seed;
 
 extern IMPBASEEXPORT std::string exe_name;
 extern IMPBASEEXPORT std::string exe_usage;
 extern IMPBASEEXPORT std::string exe_description;
 
-extern IMPBASEEXPORT boost::int64_t number_of_threads;
+#ifdef _OPENMP
+#define IMP_BASE_HAS_OPENMP 1
+#else
+#define IMP_BASE_HAS_OPENMP 0
+#endif
+
+extern IMPBASEEXPORT AdvancedFlag<boost::int64_t, IMP_BASE_HAS_OPENMP>
+    number_of_threads;
 
 extern IMPBASEEXPORT boost::scoped_ptr<boost::progress_display> progress;
 
-extern IMPBASEEXPORT bool print_deprecation_messages;
-extern IMPBASEEXPORT bool exceptions_on_deprecation;
-extern IMPBASEEXPORT base::set<std::string> printed_deprecation_messages;
+extern IMPBASEEXPORT AdvancedFlag<bool> no_print_deprecation_messages;
+extern IMPBASEEXPORT AdvancedFlag<bool> exceptions_on_deprecation;
+extern IMPBASEEXPORT boost::unordered_set<std::string>
+    printed_deprecation_messages;
 
 struct Timing {
   double total_time;
   unsigned int calls;
   Timing() : total_time(0), calls(0) {}
 };
-extern IMPBASEEXPORT base::map<std::string, Timing> timings;
+extern IMPBASEEXPORT boost::unordered_map<std::string, Timing> timings;
 IMPBASE_END_INTERNAL_NAMESPACE
 
 #endif /* IMPBASE_INTERNAL_STATIC_H */